Create main.py
Note: Twilio wouldn't allow me to send a text message due to authentication reasons, so I cannot confirm that aspect of the code is working. However, the logic seems to be correct, and error messages on Twilio's site indicate the messages would send properly if I was able to authenticate properly.